POP UP Condition
Condition de la pop up Yes, No Terminer

+++ src/vue/VuePays.java
@@ -181,19 +181,26 @@ public class VuePays extends JPanel implements ActionListener
         }
         else if (e.getSource() == this.btSupprimer)
         {
-            JOptionPane.showMessageDialog(this, "Vous risquez d'endommager la base de donnée, vérifiez qu'aucun(e) athlete (ou equipe) n'est lier a ce pays !", "!!!!!! Attention !!!!!!", JOptionPane.WARNING_MESSAGE);
-            int idPays = Integer.parseInt(txtId.getText());
-            Pays unPays = new Pays(idPays, txtLibelle.getText(), txtDescription.getText(), txtImage.getText());
+            int dialogButton = JOptionPane.YES_NO_OPTION;
+            int dialogResult = JOptionPane.showConfirmDialog (null, "Vous risquez d'endommager la base de données, \n vérifiez qu'aucun(e) athlete (ou sport) n'est lier a ce pays ! \n Voulez-vous continuer ?","Warning",dialogButton);
+            if(dialogResult == JOptionPane.NO_OPTION)
+            {
+                System.out.println("Rien n'a ete suppr");
+            }
+            else {
+                int idPays = Integer.parseInt(txtId.getText());
+                Pays unPays = new Pays(idPays, txtLibelle.getText(), txtDescription.getText(), txtImage.getText());
 
-            Modele.deletePays(unPays);
-            JOptionPane.showMessageDialog(this, "Supression effectuée");
-            txtId.setText("");
-            txtLibelle.setText("");
-            txtImage.setText("");
-            txtDescription.setText("");
+                Modele.deletePays(unPays);
+                JOptionPane.showMessageDialog(this, "Supression effectuée");
+                txtId.setText("");
+                txtLibelle.setText("");
+                txtImage.setText("");
+                txtDescription.setText("");
 
-            int rowIndex = tablePays.getSelectedRow();
-            unTableau.remove(rowIndex);
+                int rowIndex = tablePays.getSelectedRow();
+                unTableau.remove(rowIndex);
+            }
         }
         else if (e.getSource() == this.btMiseAJour)
         {
